Reverted to using the caching storage, to make sure this was the reason for the performance differences. Benchmark calculates scalability correctly now.

+++ src/main/scala/signalcollect/benchmark/Benchmark.scala
@@ -85,7 +85,7 @@ object Benchmark extends App {
   evalGraph.setCollectThreshold(0.0)
   stats = evalGraph.execute
   val computationTime6Workers = stats.computationTimeInMilliseconds
-  val scalabilityScore = ((computationTime3Workers.get / computationTime6Workers.get) * 100.0) - 100.0
+  val scalabilityScore = ((computationTime3Workers.get.toDouble / computationTime6Workers.get.toDouble) * 100.0) - 100.0
   println("Scalability Score:\t" + scalabilityScore.toInt + "%")
   evalGraph.shutDown
